test(select-popover): migrate
