+++ core/src/components/select-popover/test/fixtures.ts
@@ -1,4 +1,4 @@
-import type { E2EPage, E2ELocator, EventSpy } from '@utils/test/playwright';
+import type { E2EPage, E2ELocator, EventSpy, E2EPageOptions } from '@utils/test/playwright';
 
 import type { SelectPopoverOption } from '../select-popover-interface';
 
@@ -18,10 +18,11 @@ export class SelectPopoverPage {
     this.page = page;
   }
 
-  async setup(options: SelectPopoverOption[], multiple = false) {
+  async setup(config: E2EPageOptions, options: SelectPopoverOption[], multiple = false) {
     const { page } = this;
 
-    await page.setContent(`
+    await page.setContent(
+      `
     <ion-popover>
       <ion-select-popover></ion-select-popover>
     </ion-popover>
@@ -30,7 +31,9 @@ export class SelectPopoverPage {
       selectPopover.options = ${JSON.stringify(options)};
       selectPopover.multiple = ${multiple};
     </script>
-    `);
+    `,
+      config
+    );
 
     const ionPopoverDidPresent = await page.spyOnEvent('ionPopoverDidPresent');
     this.ionPopoverDidDismiss = await page.spyOnEvent('ionPopoverDidDismiss');
